third_party/blink/web_tests/platform/generic/fast/css/border-image-style-length-expected.txt

Test the style.length of an element with a border-image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

border-image
PASS testElement.style.length is 5
PASS testElement.style.length is 5
PASS testElement.style.length is 5
PASS testElement.style.length is 0
-webkit-border-image
PASS testElement.style.length is 1
PASS testElement.style.length is 1
PASS testElement.style.length is 1
PASS testElement.style.length is 0
-webkit-mask-box-image
PASS testElement.style.length is 5
PASS testElement.style.length is 5
PASS testElement.style.length is 5
PASS testElement.style.length is 0
PASS successfullyParsed is true

TEST COMPLETE
